refactor : Participant 인터페이스에서 Participant 추상 클래스로 수정, Name 클래스에 getter와 trim() 추가
