+++ src/main/java/blackjack/domain/participants/Dealer.java
@@ -1,34 +1,19 @@
 package blackjack.domain.participants;
 
-import java.util.Objects;
-
-import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
 
-public class Dealer implements Participant {
+public class Dealer extends Participant {
     public static final int DEALER_DRAW_CRITERIA = 17;
-
-    private Hand hand;
+    private static final String DEALER_NAME = "딜러";
 
     public Dealer() {
-        this.hand = new Hand();
+        super(DEALER_NAME);
     }
 
     public int countAddedCard() {
         return hand.countAddedCard();
     }
 
-    @Override
-    public void draw(Card card) {
-        hand.add(card);
-    }
-
-    @Override
-    public int score() {
-        return hand.calculate();
-    }
-
-    @Override
     public void drawMoreCard(final Deck deck) {
         while (needsMoreCard()) {
             draw(deck.pop());
@@ -37,41 +22,5 @@ public class Dealer implements Participant {
 
     public boolean needsMoreCard() {
         return hand.calculate() < DEALER_DRAW_CRITERIA;
-    }
-
-    @Override
-    public boolean isDealer() {
-        return true;
-    }
-
-    @Override
-    public String handStatus() {
-        return hand.toString();
-    }
-
-    @Override
-    public String getName() {
-        return "딜러";
-    }
-
-    @Override
-    public int countHand() {
-        return hand.size();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        Dealer dealer = (Dealer)o;
-        return Objects.equals(hand, dealer.hand);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(hand);
     }
 }
